fix: remove Expression::Date and cast iterations to usize
- Remove Expression::Date variant usage (not in AST)
- Cast sim.iterations (u32) to usize for executor.iterations
- Partial fix for type mismatches

+++ src/executor.rs
@@ -174,7 +174,7 @@ impl Executor {
         // Extract iteration count
         for stmt in &program.statements {
             if let Statement::Simulate(sim) = stmt {
-                self.iterations = sim.iterations;
+                self.iterations = sim.iterations as usize;
                 break;
             }
         }
